convert getWorkoutAsync -> getWorkoutOptionsAsync
