+++ src/features/exercise/thunks/index.ts
@@ -1,4 +1,4 @@
-export { default as getWorkoutAsync } from './getWorkoutAsync'
+export { default as getWorkoutOptionsAsync } from './getWorkoutOptionsAsync'
 export { default as getMorePreviousWorkouts } from './getMorePreviousWorkouts'
 export { default as getExerciseTemplates } from './getExerciseTemplates'
 export {default as postExerciseEntries} from './postExerciseEntries'
